requirements.txt
Flask==2.3.2
kafka-python==2.0.2
transformers==4.34.0
torch==2.0.1
joblib==1.3.2
pymongo==4.5.0
